# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhatsApp AI Business Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses:
    get:
      operationId: listBusinesses
      tags: [businesses]
      summary: List all businesses
      responses:
        "200":
          description: Business list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Business"
    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Create a business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessBody"
      responses:
        "201":
          description: Created business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/{id}:
    get:
      operationId: getBusiness
      tags: [businesses]
      summary: Get a business by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: Not found
    patch:
      operationId: updateBusiness
      tags: [businesses]
      summary: Update a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessBody"
      responses:
        "200":
          description: Updated business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: Not found
    delete:
      operationId: deleteBusiness
      tags: [businesses]
      summary: Delete a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Customers
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings
      parameters:
        - $ref: "#/components/parameters/BusinessIdParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, completed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Booking list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingListResponse"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      parameters:
        - $ref: "#/components/parameters/BusinessIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Approve, reject, or complete a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/BusinessIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    delete:
      operationId: deleteBooking
      tags: [bookings]
      summary: Delete a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/BusinessIdParam"
      responses:
        "204":
          description: Deleted

  # WhatsApp Webhook
...
components:
  parameters:
    BusinessIdParam:
      name: businessId
      in: query
      description: Business ID to scope the request to. Defaults to the first business if not provided.
      schema:
        type: integer
  schemas:
...
    Business:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        whatsappPhoneNumberId:
          type: ["string", "null"]
        whatsappAccessToken:
          type: ["string", "null"]
        verifyToken:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - whatsappPhoneNumberId
        - whatsappAccessToken
        - verifyToken
        - createdAt
        - updatedAt
    CreateBusinessBody:
      type: object
      properties:
        name:
          type: string
        whatsappPhoneNumberId:
          type: string
        whatsappAccessToken:
          type: string
        verifyToken:
          type: string
      required:
        - name
    UpdateBusinessBody:
      type: object
      properties:
        name:
          type: string
        whatsappPhoneNumberId:
          type: string
        whatsappAccessToken:
          type: string
        verifyToken:
          type: string
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerPhone:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        service:
          type: ["string", "null"]
        requestedDate:
          type: ["string", "null"]
        requestedTime:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, completed]
        rating:
          type: ["integer", "null"]
        ratingAskedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - customerPhone
        - customerName
        - service
        - requestedDate
        - requestedTime
        - notes
        - status
        - rating
        - ratingAskedAt
        - createdAt
        - updatedAt
    CreateBookingBody:
      type: object
      properties:
        customerId:
          type: integer
        service:
          type: string
        requestedDate:
          type: string
        requestedTime:
          type: string
        notes:
          type: string
      required:
        - customerId
    UpdateBookingBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected, completed]
        notes:
          type: string
    BookingListResponse:
      type: object
      properties:
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - bookings
        - total
        - page
        - limit
...